requirements.txt
# Flask web framework
Flask==2.3.3
Werkzeug==2.3.7

# Environment variables
python-dotenv==1.0.0

# Machine Learning & Computer Vision
torch==1.13.1
torchvision==0.14.1
Pillow==9.0.0
numpy==1.24.4

# Google Cloud Dialogflow
google-cloud-dialogflow==2.25.0

